requirements.txt
torch
torchvision
torchfcn
numpy
pillow